extras: fix LZ4+ph split decoder — uint16_t-cast bug truncated bounds past 64 KB
The "WIP, decoder buggy at 64-KB boundary" from the prior commit was
actually a one-line bug in lz4_split_decode.c, not in the modified
LZ4 encoder.  The offset-bounds check

    if (offset == 0 || offset > (uint16_t)(out_p - out)) return -4;

cast `out_p - out` to uint16_t, which wraps past 65535.  For any
valid LZ4 offset on output positions > 64 KB, the cast made the
check spuriously fail, the decoder bailed out, and bench's
memcmp-based verification reported "split decode mismatch" right at
the wrap point.

Fix: compare in size_t.

    if (offset == 0 || (size_t)offset > (size_t)(out_p - out)) return -4;

The walker-vs-encoder content comparison added in this commit
verifies the encoder is producing the correct 4 streams byte-for-byte
against a parse of the standard LZ4 output — useful regression
detector for any future LZ4-source-side hacks.

Roundtrip now passes on every MAIN dataset.  Final sweep saved to
results/lz4_ph-m4-20260517-2020-*.txt.  Headline numbers (M4 Max,
decode MB/s = source bytes / total decode time):

  dataset       LZ4 alone        LZ4+ph         zstd             LZ4+split (NEW)
                ratio  MB/s     ratio  MB/s     ratio  MB/s     ratio  MB/s
  cat-wiki.html 0.214  7488     0.197  5782     0.159  3098     0.185  1444
  pride.txt     0.411  4894     0.388  3046     0.321  1613     0.355   742
  json_api      0.152 17242     0.139  7484     0.111  5092     0.138  2468
  chinese_text  0.531  4458     0.510  2696     0.435  1403     0.478   681
  calgary_pic   0.130 20999     0.117  8334     0.097  4644     0.113  1999
  log_apache    0.111 22222     0.103  9542     0.077  6876     0.108  3173
  source_c      0.243  8336     0.222  4217     0.191  2714     0.226  1089
  dna_fasta     0.400  4914     0.362  2864     0.304  1323     0.326   847
  csv_numeric   0.320  6619     0.288  3673     0.227  1980     0.261  1060
  cat-image.jpg 0.965 36312     0.965  3160     0.963 15246     0.998  1820
  gzip_random   1.003 46340     1.044  2693     1.000 73466     1.084  2376

Key reads on the split codec:

- Compression: semantic-split ph closes ~50-70% of the LZ4+ph → zstd
  gap on text-heavy data (cat-wiki 0.197 → 0.185 → 0.159 = 63%
  closed; pride 0.388 → 0.355 → 0.321 = 49%; log_apache 0.103 →
  0.108 → 0.077 = NEGATIVE).
- Hurts on already-incompressible inputs (gzip_random, cat-image):
  the 4 per-stream pivcohuf headers + per-stream Huffman tables cost
  more than the per-stream entropy savings when streams have no skew.
- Decode speed: ~3-4× slower than LZ4+ph stacked.  Dominated by 4
  separate pivcohuf_decompress calls (each pays pivcohuf header
  parse + Huffman table rebuild + 8-KB-block decode loop) — for
  small sections (overflow, sometimes literals) the per-call
  overhead dwarfs the work.  Real LZ4+ph-split would want a leaner
  single-block ph encoder or merge ph-decode into the lz4-split-
  decoder's inner loop.  Out of scope for the prototype.

Net Pareto position: split codec is between LZ4+ph and zstd on
ratio, BELOW both on decode speed.  Not a Pareto win in this
implementation — confirms that the SIZE benefit of semantic split
is real (matches the earlier projection) but the SPEED cost of 4
separate ph passes wipes out ph's decode-throughput advantage.
Practical takeaway: if you want compression closer to zstd, do the
work to merge the 4 ph passes into one fused decoder; otherwise stack
ph on the full LZ4 stream (the LZ4+ph column).

+++ extras/lz4_split_decode.c
@@ -79,7 +79,12 @@ int lz4_split_decompress(const uint8_t *literals, size_t literals_len,
         if (off_p + 2 > offsets + offsets_len) return -3;
         uint16_t offset = (uint16_t)off_p[0] | ((uint16_t)off_p[1] << 8);
         off_p += 2;
-        if (offset == 0 || offset > (uint16_t)(out_p - out)) return -4;
+        /* Offset bounds check: must be in (0, output-so-far].  Note:
+         * `offset` is u16 (max 65535) but `out_p - out` can be much
+         * larger than 64 KB on long files — a naïve `(uint16_t)`
+         * cast truncates and produces false positives past the
+         * 64-KB boundary.  Compare in size_t. */
+        if (offset == 0 || (size_t)offset > (size_t)(out_p - out)) return -4;
 
         /* Decode match length (token lo-nibble + 4, plus overflow). */
         size_t match_len = (size_t)(token & 0xf);
